centralize format_address
